Implement a basic parser for function expressions

+++ src/lexer/literal_table.rs
@@ -6,14 +6,14 @@ pub struct LiteralTable {
 }
 
 impl LiteralTable {
-    pub fn insert(&mut self, literal: Literal) -> usize {
-        let index = self.literals.len();
+    pub fn insert(&mut self, literal: Literal) -> u32 {
+        let index = self.literals.len() as u32;
         self.literals.push(literal);
         index
     }
 
-    pub fn get(&self, index: usize) -> Option<&Literal> {
-        self.literals.get(index)
+    pub fn get(&self, index: u32) -> Option<&Literal> {
+        self.literals.get(index as usize)
     }
 
     pub fn iter(&self) -> impl Iterator<Item = &Literal> {
